后台借款人添加BUG修复。 添加返回主键，model直接调用get id

+++ com.jebao.root/2.2-com.jebao.erp.service.inf/src/main/java/com/jebao/erp/service/inf/loaner/ILoanerServiceInf.java
@@ -24,7 +24,7 @@ public interface ILoanerServiceInf {
      * @param entity
      * @return
      */
-    Long addLoaner(TbLoaner entity);
+    int addLoaner(TbLoaner entity);
 
     /**
      * 根据手机号查询借款人基本信息
